+++ docs/source/changelog.rst
@@ -11,7 +11,7 @@ New Features
 - Add install guidelines for different platforms (`pixi`, `conda`, `pip`, `uv`) in docs (:gh:`1108` by `Julian Tachella`_)
 - Add :class:`deepinv.optim.SIRT` algorithm for tomographic reconstruction (:gh:`985` by `Thibaut Modrzyk`_`)
 - Add :class:`deepinv.optim.MLEM` algorithm for Poisson inverse problems (:gh:`1051` by `Thibaut Modrzyk`_)
-- Add :class:`deepinv.loss.L2RLoss` loss for self-supervised denoising learning with unknown noise (:gh:`1150` by `Brayan Monroy`_)
+- Add :class:`deepinv.loss.Learning2RecorruptLoss` loss for self-supervised denoising learning with unknown noise (:gh:`1150` by `Brayan Monroy`_)
 - Add the equivariant splitting loss :class:`deepinv.loss.EquivariantSplittingLoss` with equivariant reconstructors :class:`deepinv.models.EquivariantReconstructor` and virtual physics :class:`deepinv.physics.VirtualLinearPhysics` (:gh:`881` by `Jérémy Scanvic`_)
 
 Changed